tests/faccessat.c: print syscall() result using %ld format

+++ tests/faccessat.c
@@ -13,10 +13,9 @@ main(void)
 	static const char sample[] = "faccessat.sample";
 	const long int fd = (long int) 0xdeadbeefffffffff;
 
-	int rc = syscall(__NR_faccessat, fd, sample, F_OK);
-	printf("faccessat(%d, \"%s\", F_OK) = %d %s (%m)\n",
-	       (int) fd, sample, rc,
-	       errno2name());
+	long rc = syscall(__NR_faccessat, fd, sample, F_OK);
+	printf("faccessat(%d, \"%s\", F_OK) = %ld %s (%m)\n",
+	       (int) fd, sample, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
